Changes required to compile on Windows (with some no-no's)
Some AudioClients allocate variable length arrays in their process or render fuction.  This was previously done as local stack variables because GCC supports VLAs from the C99 standard.  Sadly MSVC does not support this.  The fix for now is to convert all these allocations to heap operations.

The other no-no is I added back locked synchronization to SampleAccumulator::get().  This function was written to do double-buffering to avoid locks, but it apparently doesn't work on Windows.  For now I just addded calls to AudioServer::EnterLock / ExitLock to avoid collisions with the audio thread.